requirements.txt
aiohttp==3.8.6 ; python_version >= "3.10" and python_version < "3.13"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "3.13"
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "3.13"
anyio==3.7.1 ; python_version >= "3.10" and python_version < "3.13"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "3.13"
asyncio==3.4.3 ; python_version >= "3.10" and python_version < "3.13"
atro-args==0.6.12 ; python_version >= "3.10" and python_version < "3.13"
atro-utils==0.0.8 ; python_version >= "3.10" and python_version < "3.13"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "3.13"
backoff==2.2.1 ; python_version >= "3.10" and python_version < "3.13"
beautifulsoup4==4.12.2 ; python_version >= "3.10" and python_version < "3.13"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "3.13"
chardet==5.2.0 ; python_version >= "3.10" and python_version < "3.13"
charset-normalizer==3.3.1 ; python_version >= "3.10" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.10" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.13" and platform_system == "Windows"
dataclasses-json==0.6.1 ; python_version >= "3.10" and python_version < "3.13"
diskcache==5.6.3 ; python_version >= "3.10" and python_version < "3.13"
emoji==2.8.0 ; python_version >= "3.10" and python_version < "3.13"
exceptiongroup==1.1.3 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.103.2 ; python_version >= "3.10" and python_version < "3.13"
filelock==3.13.1 ; python_version >= "3.10" and python_version < "3.13"
filetype==1.2.0 ; python_version >= "3.10" and python_version < "3.13"
frozenlist==1.4.0 ; python_version >= "3.10" and python_version < "3.13"
fsspec==2023.10.0 ; python_version >= "3.10" and python_version < "3.13"
greenlet==3.0.1 ; python_version >= "3.10" and python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
grpcio-tools==1.59.2 ; python_version >= "3.10" and python_version < "3.13"
grpcio==1.59.2 ; python_version >= "3.10" and python_version < "3.13"
h11==0.14.0 ; python_version >= "3.10" and python_version < "3.13"
h2==4.1.0 ; python_version >= "3.10" and python_version < "3.13"
hpack==4.0.0 ; python_version >= "3.10" and python_version < "3.13"
httpcore==0.18.0 ; python_version >= "3.10" and python_version < "3.13"
httpx[http2]==0.25.0 ; python_version >= "3.10" and python_version < "3.13"
huggingface-hub==0.17.3 ; python_version >= "3.10" and python_version < "3.13"
hyperframe==6.0.1 ; python_version >= "3.10" and python_version < "3.13"
idna==3.4 ; python_version >= "3.10" and python_version < "3.13"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "3.13"
joblib==1.3.2 ; python_version >= "3.10" and python_version < "3.13"
jsonpatch==1.33 ; python_version >= "3.10" and python_version < "3.13"
jsonpointer==2.4 ; python_version >= "3.10" and python_version < "3.13"
langchain==0.0.314 ; python_version >= "3.10" and python_version < "3.13"
langchainhub==0.1.13 ; python_version >= "3.10" and python_version < "3.13"
langdetect==1.0.9 ; python_version >= "3.10" and python_version < "3.13"
langsmith==0.0.54 ; python_version >= "3.10" and python_version < "3.13"
llama-cpp-python==0.2.11 ; python_version >= "3.10" and python_version < "3.13"
lxml==4.9.3 ; python_version >= "3.10" and python_version < "3.13"
markdown==3.5 ; python_version >= "3.10" and python_version < "3.13"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "3.13"
marshmallow==3.20.1 ; python_version >= "3.10" and python_version < "3.13"
mpmath==1.3.0 ; python_version >= "3.10" and python_version < "3.13"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "3.13"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "3.13"
networkx==3.2.1 ; python_version >= "3.10" and python_version < "3.13"
nltk==3.8.1 ; python_version >= "3.10" and python_version < "3.13"
numpy==1.26.1 ; python_version >= "3.10" and python_version < "3.13"
openai==0.28.1 ; python_version >= "3.10" and python_version < "3.13"
packaging==23.2 ; python_version >= "3.10" and python_version < "3.13"
pillow==10.1.0 ; python_version >= "3.10" and python_version < "3.13"
portalocker==2.8.2 ; python_version >= "3.10" and python_version < "3.13"
protobuf==4.24.4 ; python_version >= "3.10" and python_version < "3.13"
pydantic-core==2.10.1 ; python_version >= "3.10" and python_version < "3.13"
pydantic==2.4.2 ; python_version >= "3.10" and python_version < "3.13"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "3.13"
python-iso639==2023.6.15 ; python_version >= "3.10" and python_version < "3.13"
python-magic==0.4.27 ; python_version >= "3.10" and python_version < "3.13"
pywin32==306 ; python_version >= "3.10" and python_version < "3.13" and platform_system == "Windows"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.13"
qdrant-client==1.6.4 ; python_version >= "3.10" and python_version < "3.13"
rapidfuzz==3.4.0 ; python_version >= "3.10" and python_version < "3.13"
regex==2023.10.3 ; python_version >= "3.10" and python_version < "3.13"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.13"
safetensors==0.4.0 ; python_version >= "3.10" and python_version < "3.13"
scikit-learn==1.3.2 ; python_version >= "3.10" and python_version < "3.13"
scipy==1.11.3 ; python_version >= "3.10" and python_version < "3.13"
sentence-transformers==2.2.2 ; python_version >= "3.10" and python_version < "3.13"
sentencepiece==0.1.99 ; python_version >= "3.10" and python_version < "3.13"
setuptools==68.2.2 ; python_version >= "3.10" and python_version < "3.13"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.13"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "3.13"
soupsieve==2.5 ; python_version >= "3.10" and python_version < "3.13"
sqlalchemy==2.0.22 ; python_version >= "3.10" and python_version < "3.13"
starlette==0.27.0 ; python_version >= "3.10" and python_version < "3.13"
sympy==1.12 ; python_version >= "3.10" and python_version < "3.13"
tabulate==0.9.0 ; python_version >= "3.10" and python_version < "3.13"
tenacity==8.2.3 ; python_version >= "3.10" and python_version < "3.13"
threadpoolctl==3.2.0 ; python_version >= "3.10" and python_version < "3.13"
tokenizers==0.14.1 ; python_version >= "3.10" and python_version < "3.13"
torch==2.1.0 ; python_version >= "3.10" and python_version < "3.13"
torchvision==0.16.0 ; python_version >= "3.10" and python_version < "3.13"
tqdm==4.66.1 ; python_version >= "3.10" and python_version < "3.13"
transformers==4.34.1 ; python_version >= "3.10" and python_version < "3.13"
types-pyyaml==6.0.12.12 ; python_version >= "3.10" and python_version < "3.13"
types-requests==2.31.0.6 ; python_version >= "3.10" and python_version < "3.13"
types-urllib3==1.26.25.14 ; python_version >= "3.10" and python_version < "3.13"
typing-extensions==4.8.0 ; python_version >= "3.10" and python_version < "3.13"
typing-inspect==0.9.0 ; python_version >= "3.10" and python_version < "3.13"
unstructured==0.10.28 ; python_version >= "3.10" and python_version < "3.13"
urllib3==1.26.18 ; python_version >= "3.10" and python_version < "3.13"
uvicorn==0.23.2 ; python_version >= "3.10" and python_version < "3.13"
yarl==1.9.2 ; python_version >= "3.10" and python_version < "3.13"
